资金float化初步完成，待深度测试

+++ ROOT_demo/Assets/Script/_Common/EventRelated/EventInfo.cs
@@ -27,10 +27,10 @@ namespace ROOT.Message
     
     public class CurrencyUpdatedInfo : RootMessageBase
     {
-        public int TotalIncomesVal = int.MaxValue;
-        public int BaseIncomesVal = int.MaxValue;
-        public int BonusIncomesVal = int.MaxValue;
-        public int CurrencyVal = int.MaxValue;
+        public float TotalIncomesVal = int.MaxValue;
+        public float BaseIncomesVal = int.MaxValue;
+        public float BonusIncomesVal = int.MaxValue;
+        public float CurrencyVal = int.MaxValue;
         public override string Type => WorldEvent.CurrencyUpdatedEvent;
     }
